requirements.txt
beautifulsoup4
playwright
pyyaml
requests
pymongo
python-dotenv
langchain-google-genai
langchain-openai
langchain
pydantic
fastapi
uvicorn